// NEST_MS/Launcher/orders-ms/src/interfaces/order.interface.ts
import { OrderStatus } from "@prisma/client";


export interface OrderWithProducts {
    orderItem: {
        name: any;
        productId: number;
        quantity: number;
        price: number;
    }[];
    id: string;
    totalAmount: number;
    totalItems: number;
    status: OrderStatus;
    paid: boolean;
    paidAt: Date | null;
    createdAt: Date;
    updatedAt: Date;
}